+++ packages/koin_flutter/test/classes.dart
@@ -1,0 +1,5 @@
+class Component {
+  const Component(this.id);
+
+  final int id;
+}