+++ packages/koin_flutter/test/keys.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/widgets.dart';
+
+final buttonKey = UniqueKey();
+final firstPageKey = UniqueKey();
+final secondPageKey = UniqueKey();