+++ packages/koin_flutter/test/scope_observer_test.dart
@@ -1,14 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:koin/koin.dart';
+import 'package:koin_flutter/internals.dart';
 import 'package:koin_flutter/koin_flutter.dart';
 import 'package:koin_test/koin_test.dart';
-
 import 'classes.dart';
 import 'keys.dart';
 
 final scopeProviderModule = Module()
   ..scope<FirstPage>((dsl) {
+    dsl.scoped((scope) => Component(10));
+  })
+  ..scope<SecondPage>((dsl) {
     dsl.scoped((scope) => Component(10));
   });
 
@@ -17,14 +20,14 @@ class FirstPage extends StatefulWidget {
   _FirstPageState createState() => _FirstPageState();
 }
 
-class _FirstPageState extends State<FirstPage> {
+class _FirstPageState extends State<FirstPage> with ScopeStateMixin {
   @override
   Widget build(BuildContext context) {
     final bloc = currentScope.get<Component>();
     return Scaffold(
         key: firstPageKey,
         appBar: AppBar(
-          title: Text("Home Page"),
+          title: Text("First Page"),
         ),
         body: Column(
           children: [
@@ -35,8 +38,7 @@ class _FirstPageState extends State<FirstPage> {
               key: buttonKey,
               onPressed: () {
                 Navigator.push(context, MaterialPageRoute(builder: (context) {
-                  return ScopeProvider(
-                      child: SecondPage(), scope: currentScope);
+                  return SecondPage();
                 }));
               },
               child: Text("Push Second Page"),
@@ -46,10 +48,15 @@ class _FirstPageState extends State<FirstPage> {
   }
 }
 
-class SecondPage extends StatelessWidget {
+class SecondPage extends StatefulWidget {
+  @override
+  _SecondPageState createState() => _SecondPageState();
+}
+
+class _SecondPageState extends State<SecondPage> with ScopeStateMixin {
   @override
   Widget build(BuildContext context) {
-    final component = context.scope.get<Component>();
+    final component = currentScope.get<Component>();
 
     return Scaffold(
       key: secondPageKey,
@@ -67,31 +74,26 @@ void main() {
 
   koinTearDown();
 
-  testWidgets('shoud get the scope passed to the context', (tester) async {
+  testWidgets('shoud observer the escope', (tester) async {
     await tester.pumpWidget(MaterialApp(
       home: FirstPage(),
     ));
 
-    final fisrtPageFinder = find.byKey(firstPageKey);
+    final firstPageFinder = find.byKey(firstPageKey);
     final secondPageFinder = find.byKey(secondPageKey);
     final buttonFinder = find.byType(FlatButton);
     final textFinder = find.text('10');
 
-    expect(fisrtPageFinder, findsOneWidget);
-
+    expect(firstPageFinder, findsOneWidget);
+    expect(
+        FlutterKoinScopeObserver.scopeWidgetObserver.scopeContexts.length, 1);
     await tester.tap(buttonFinder);
     await tester.pumpAndSettle();
 
     expect(secondPageFinder, findsOneWidget);
+    expect(
+        FlutterKoinScopeObserver.scopeWidgetObserver.scopeContexts.length, 2);
 
     expect(textFinder, findsOneWidget);
-  });
-  testWidgets('shoud trow a exception when the context does not have a scope',
-      (tester) async {
-    await tester.pumpWidget(MaterialApp(
-      home: SecondPage(),
-    ));
-
-    expect(tester.takeException(), isInstanceOf<ScopeNotFoundException>());
   });
 }